Prefer versions on disk in getExecutable
Example:

    getExecutable({
      name: "elm-json",
      version: "~0.2.8",
      onProgress: (percentage) => {
        console.log(percentage);
      },
    })

And ~/.elm/elm-tooling/elm-json/0.2.8 exists.

Previously, if elm-tooling has 0.2.10 in KNOWN_TOOLS, `getExecutable`
would resolve to 0.2.10 – the latest known version matching the version
range.

Now, `getExecutable` would resolve to 0.2.8 – the latest version on disk
that matches the version range (and is a version we recognize).

If a user has `elm-json` in their elm-tooling.json, this increases the
likelyhood that packages that use `getExecutable` don’t download a
second `elm-json` (but a newer version).